+++ lib/paranoia.rb
@@ -62,27 +62,15 @@ module Paranoia
   end
 
   def destroy
-    callbacks_result = transaction do
+    transaction do
       run_callbacks(:destroy) do
         touch_paranoia_column unless destroyed?
       end
     end
-    callbacks_result ? self : false
-  end
-
-  # As of Rails 4.1.0 +destroy!+ will no longer remove the record from the db
-  # unless you touch the paranoia column before.
-  # We need to override it here otherwise children records might be removed
-  # when they shouldn't
-  if ActiveRecord::VERSION::STRING >= "4.1"
-    def destroy!
-      destroyed? ? super : destroy || raise(ActiveRecord::RecordNotDestroyed)
-    end
   end
 
   def delete
-    return if new_record?
-    touch_paranoia_column(false)
+    touch_paranoia_column
   end
 
   def restore!(opts = {})
@@ -107,10 +95,10 @@ module Paranoia
   end
   alias :restore :restore!
 
-  def destroyed?
+  def paranoia_destroyed?
     send(paranoia_column) != paranoia_sentinel_value
   end
-  alias :deleted? :destroyed?
+  alias :deleted? :paranoia_destroyed?
 
   private
 
@@ -118,16 +106,13 @@ module Paranoia
   # insert time to paranoia column.
   # @param with_transaction [Boolean] exec with ActiveRecord Transactions.
   def touch_paranoia_column(with_transaction=false)
-    # This method is (potentially) called from really_destroy
-    # The object the method is being called on may be frozen
-    # Let's not touch it if it's frozen.
-    unless self.frozen?
-      if with_transaction
-        with_transaction_returning_status { touch(paranoia_column) }
-      else
-        touch(paranoia_column)
-      end
+    raise ActiveRecord::ReadOnlyRecord, "#{self.class} is marked as readonly" if readonly?
+    if persisted?
+      touch(paranoia_column)
+    elsif !frozen?
+      write_attribute(paranoia_column, current_time_from_proper_timezone)
     end
+    self
   end
 
   # restore associated records that have been soft deleted when
@@ -154,17 +139,20 @@ module Paranoia
       end
 
       if association_data.nil? && association.macro.to_s == "has_one"
-        association_class_name = association.options[:class_name].present? ? association.options[:class_name] : association.name.to_s.camelize
-        association_foreign_key = association.foreign_key.present? ? association.foreign_key : "#{self.class.name.to_s.underscore}_id"
+        association_class_name = association.class_name
+        association_foreign_key = association.foreign_key
 
         if association.type
           association_polymorphic_type = association.type
-          where_conditions = { association_polymorphic_type => self.class.name.to_s, association_foreign_key => self.id }
+          association_find_conditions = { association_polymorphic_type => self.class.name.to_s, association_foreign_key => self.id }
         else
-          where_conditions = { association_foreign_key => self.id }
+          association_find_conditions = { association_foreign_key => self.id }
         end
 
-        Object.const_get(association_class_name).only_deleted.where(where_conditions).first.try(:restore, recursive: true)
+        association_class = Object.const_get(association_class_name)
+        if association_class.paranoid?
+          association_class.only_deleted.where(association_find_conditions).first.try!(:restore, recursive: true)
+        end
       end
     end
 
@@ -174,27 +162,30 @@ end
 
 class ActiveRecord::Base
   def self.acts_as_paranoid(options={})
-    alias :destroy! :destroy
-    alias :delete! :delete
+    alias :really_destroyed? :destroyed?
+    alias :really_delete :delete
+
+    alias :destroy_without_paranoia :destroy
     def really_destroy!
       dependent_reflections = self.class.reflections.select do |name, reflection|
         reflection.options[:dependent] == :destroy
       end
       if dependent_reflections.any?
-        dependent_reflections.each do |name, _|
-          associated_records = self.send(name)
+        dependent_reflections.each do |name, reflection|
+          association_data = self.send(name)
           # has_one association can return nil
-          if associated_records && associated_records.respond_to?(:with_deleted)
-            # Paranoid models will have this method, non-paranoid models will not
-            associated_records.with_deleted.each(&:really_destroy!)
-            self.send(name).reload
-          elsif associated_records && !associated_records.respond_to?(:each) # single record
-            associated_records.really_destroy!
+          # .paranoid? will work for both instances and classes
+          if association_data && association_data.paranoid?
+            if reflection.collection?
+              association_data.with_deleted.each(&:really_destroy!)
+            else
+              association_data.really_destroy!
+            end
           end
         end
       end
-      touch_paranoia_column if ActiveRecord::VERSION::STRING >= "4.1"
-      destroy!
+      write_attribute(paranoia_column, current_time_from_proper_timezone)
+      destroy_without_paranoia
     end
 
     include Paranoia
@@ -230,13 +221,6 @@ class ActiveRecord::Base
 
   def self.paranoid? ; false ; end
   def paranoid? ; self.class.paranoid? ; end
-
-  # Override the persisted method to allow for the paranoia gem.
-  # If a paranoid record is selected, then we only want to check
-  # if it's a new record, not if it is "destroyed".
-  def persisted?
-    paranoid? ? !new_record? : super
-  end
 
   private